Bulit In Commands
Job list small progress, started
External commands to make space for
Job list, works + errors.
Next is built in commands related to
job list + methods on job list.
Note: the symbol & should be ignored in
built in commands, we should check for it
and remove it form args[] in factory but keep
it in Command::cmd_line.