Built-In Commands
fg and bg commands started, errors started but not finished.
basic behavior is running as expected but testing is needed.
before moving to the next command we should clean the c'tor
from *smash since it's a singelton. and we should finish fg
and bg commands including errors and test it as well as other
commands using JobList.